Docstring, use *args an **kw.
Always write the docs first :)

I decided that since the natural way to write the args is using the
*args or **kw syntax, I'd just go ahead and allow that, reserving '_'
names for settings.  Of which there are none yet, but I want to add one
to include the parameter list name in the arguments passed to the test
function.  There could also be another to provide a "name generation"
function.