Add _include_key setting, drop other TODO.
Also throw an error if an unexpected setting name is passed in.  That
is, we're reserving the '_' namespace for settings, which shouldn't
be a problem since one would be unlikely to want to use a leading
underscore in the name of a parameter list.

I decided to drop the name creation function TODO.  I don't need it,
so lets wait for someone who does to request it.