Rename module, and allow skipping * and **.
Packages generally don't have underscores in python, and I'd forgotten
that somehow.  I thought it was causing me some trouble with pip, and
while that may have been a red herring, I've decided to keep the
non-underscore name.

I also found that in actual use, when you are using the same parameter
lists for more than one test call, it is more natural to not have
to specify the * or **.